Optimize Looped replaceAll() Calls
Several places in all versions of RndLearner were calling
replaceAll() in a loop (for each variable) instead of using the vector
version. These have been replaced to use the vector version, which
results in a noticable speed-up (especially for V1, which was using them
in the process of checking lemmas).