Fix SeedMiner Out Of Bounds Array Read
Code in SeedMiner was assuming that hr.srcVars.size() ==
hr.dstVars.size(), meaning an out-of-bounds read if the two lengths
weren't the same (as is often the case for non-transition-system CHCs).
This resulted in a Z3 failed assertion as a sort AST node (INT) appeared
in later stages of the code. This has been fixed.

To reproduce:
    freqhorn --v1 bench_horn_multiple/samples_multiple_inv_06.smt2