Initial (Hacky) Support for ANY_*_CONST
This adds support for the special non-terminal 'ANY_SORT_CONST' which
expands to any 'SORT' constant. Under the hood, this uses Z3's model
production facilities to find that constant, falling back to an
enumerative approach instead.

Support is primitive; solving is slow, only single invariants systems
are supported, and non-integer constant support is untested and likely
to crash. The implementation is sprawling and likely more complex than
it needs to be. A refactor or recommit is soon in the future; this is
mostly here to let me test it.